requirements.txt
autogen-core
autogen-agentchat
autogen-ext
python-dotenv
openai